+++ examples/arrow-function-component.jsx
@@ -1,11 +1,12 @@
 import React from 'react';
+import styles from './TestComponent.module.css';
 
 const TestComponent = (props) => {
     return (
-        <>
+        <div className={styles.root}>
             {/* Add your component content here */}
-        </>
+        </div>
     );
 }
 
-export default TestComponent; 
+export default TestComponent;